+++ Modules/Loan/database/migrations/2025_05_03_173993_create_loan_notes_table.php
@@ -11,6 +11,10 @@ return new class extends Migration
      */
     public function up(): void
     {
+        if (Schema::hasTable('loan_notes')) {
+            return;
+        }
+
         Schema::create('loan_notes', function (Blueprint $table) {
             $table->id();
             $table->foreignId('loan_id')->constrained()->onDelete('cascade');
