+++ database/migrations/2025_05_03_173732_create_custom_subscriptions_table.php
@@ -8,6 +8,10 @@ return new class extends Migration
 {
     public function up(): void
     {
+        if (Schema::hasTable('custom_subscriptions')) {
+            return;
+        }
+
         Schema::create('custom_subscriptions', function (Blueprint $table) {
             $table->id();
             $table->foreignId('user_id')->constrained()->onDelete('cascade');
